added new exceptions and minor refactoring
